Working Mandlebrot

+++ Mandlebrot_Set/ComplexPlane.h
@@ -7,15 +7,15 @@ class ComplexPlane
 {
 	private:
 		const unsigned int MAX_ITER = 64;
-		const float BASE_WIDTH = 4.0;
-		const float BASE_HEIGHT = 4.0;
-		const float BASE_ZOOM = 0.5;
+		const double BASE_WIDTH = 4.0;
+		const double BASE_HEIGHT = 4.0;
+		const double BASE_ZOOM = 0.5;
 		Vector2f m_mouseLocation;
 		View m_view;
 		int m_zoomCount;
 		float m_aspectRatio;
 	public:
-		ComplexPlane(float aspectRatio);
+		ComplexPlane(double aspectRatio);
 		void zoomIn();
 		void zoomOut();
 		void setCenter(Vector2f coord);
